Fix build when template is instantiated with double type

+++ examples/pbr_surface/main.cc
@@ -24,7 +24,7 @@
 #include "mesh.h"
 #include "utility.h"
 
-using float_precision = float;
+using float_precision = double;
 
 void loadSampler(pbr_maths::sampler2D < float_precision >& sampler,
                  const stbi_uc* data, int w,
